Move spotdl.py inside spotdl directory
