--- requirements.txt
@@ -1,12 +1,0 @@
-pathlib >= 1.0.1
-youtube_dl >= 2017.5.1
-pafy >= 0.5.3.1
-spotipy >= 2.4.4
-mutagen >= 1.37
-beautifulsoup4 >= 4.6.0
-unicode-slugify >= 0.1.3
-titlecase >= 0.10.0
-logzero >= 1.3.1
-lyricwikia >= 0.1.8
-PyYAML >= 3.12
-appdirs >= 1.4.3
