+++ spotdl/metadata.py
@@ -2,7 +2,7 @@ from mutagen.easyid3 import EasyID3
 from mutagen.id3 import ID3, TORY, TYER, TPUB, APIC, USLT, COMM
 from mutagen.mp4 import MP4, MP4Cover
 from mutagen.flac import Picture, FLAC
-from core.const import log, TAG_PRESET, M4A_TAG_PRESET
+from spotdl.const import log, TAG_PRESET, M4A_TAG_PRESET
 
 import urllib.request
