+++ spotdl/spotify_tools.py
@@ -2,8 +2,8 @@ import spotipy
 import spotipy.oauth2 as oauth2
 import lyricwikia
 
-from core import internals
-from core.const import log
+from spotdl import internals
+from spotdl.const import log
 
 from slugify import slugify
 from titlecase import titlecase
