+++ test/test_with_metadata.py
@@ -1,12 +1,12 @@
 
-from core import const
-from core import internals
-from core import spotify_tools
-from core import youtube_tools
-from core import convert
-from core import metadata
+from spotdl import const
+from spotdl import internals
+from spotdl import spotify_tools
+from spotdl import youtube_tools
+from spotdl import convert
+from spotdl import metadata
 
-import spotdl
+from spotdl import spotdl
 
 import loader
 import os
